+++ config/toml.go
@@ -162,8 +162,8 @@ seed_mode = {{ .P2P.SeedMode }}
 # Authenticated encryption
 auth_enc = {{ .P2P.AuthEnc }}
 
-# Comma separated list of nodes to keep private (will not be gossiped to other peers) connections to
-private_peers = {{ .P2P.PrivatePeers }}
+# Comma separated list of peer IDs to keep private (will not be gossiped to other peers)
+private_peer_ids = {{ .P2P.PrivatePeerIDs }}
 
 ##### mempool configuration options #####
 [mempool]