+++ CHANGELOG.md
@@ -38,7 +38,7 @@ IMPROVEMENTS:
   release after this one)
 
 FEATURES:
-- [config] added the `--p2p.private_peers` flag and `PrivatePeers` config variable (see config for description)
+- [config] added the `--p2p.private_peer_ids` flag and `PrivatePeerIDs` config variable (see config for description)
 
 ## 0.16.0 (February 20th, 2017)